Extract ReplaceValue to ServerConfigReplacer class
