+++ ArmaServerManagerTests/ServerConfigReplacerTests.cs
@@ -8,9 +8,9 @@ using Microsoft.Extensions.Configuration;
 using Xunit;
 
 namespace ArmaServerManagerTests {
-    public class ServerConfigTests {
+    public class ServerConfigReplacerTests {
         [Fact]
-        public void ServerConfig_ReplaceValue_StringEqual() {
+        public void ReplaceValue_StringEqual() {
             var modsetConfig = new ConfigurationBuilder()
                 .SetBasePath(Directory.GetCurrentDirectory())
                 .AddJsonFile("example_common.json")
@@ -19,14 +19,14 @@ namespace ArmaServerManagerTests {
             var key = "hostName"; // It has it's value as string with quotes
             var value = modsetConfig.GetSection("server")[key];
 
-            var newCfgFile = ServerConfig.ReplaceValue(cfgFile, key, value);
+            var newCfgFile = ServerConfigReplacer.ReplaceValue(cfgFile, key, value);
 
             var match = Regex.IsMatch(newCfgFile, $"{key} = \"{value}\";");
             Assert.True(match);
         }
 
         [Fact]
-        public void ServerConfig_ReplaceValue_NoQuotesEqual() {
+        public void ReplaceValue_NoQuotesEqual() {
             var modsetConfig = new ConfigurationBuilder()
                 .SetBasePath(Directory.GetCurrentDirectory())
                 .AddJsonFile("example_common.json")
@@ -35,14 +35,14 @@ namespace ArmaServerManagerTests {
             var key = "verifySignatures"; // It has it's value as number with no quotes
             var value = modsetConfig.GetSection("server")[key];
 
-            var newCfgFile = ServerConfig.ReplaceValue(cfgFile, key, value);
+            var newCfgFile = ServerConfigReplacer.ReplaceValue(cfgFile, key, value);
 
             var match = Regex.IsMatch(newCfgFile, $"{key} = {value};");
             Assert.True(match);
         }
 
         [Fact]
-        public void ServerConfig_ReplaceValue_ArrayEqual() {
+        public void ReplaceValue_ArrayEqual() {
             var modsetConfig = new ConfigurationBuilder()
                 .SetBasePath(Directory.GetCurrentDirectory())
                 .AddJsonFile("example_common.json")
@@ -52,7 +52,7 @@ namespace ArmaServerManagerTests {
             var value = modsetConfig.GetSection("server").GetSection(key).GetChildren().ToList();
             var stringValue = String.Join(", ", value.Select(p => p.Value.ToString()));
 
-            var newCfgFile = ServerConfig.ReplaceValue(cfgFile, key, stringValue);
+            var newCfgFile = ServerConfigReplacer.ReplaceValue(cfgFile, key, stringValue);
 
             key = @"admins\[\]"; // Change for regex
             var match = Regex.IsMatch(newCfgFile, $"{key} = {{{stringValue}}};");
